Moved ombi functions to app/ombi.py
