+++ app/__init__.py
@@ -111,4 +111,4 @@ database.create_tables(
 if __name__ == "__main__":
     app.run()
 
-from app import admin, web, plex, tasks, jellyfin, helpers
+from app import admin, web, plex, tasks, jellyfin, helpers, ombi
