+++ app/admin.py
@@ -1,6 +1,7 @@
 from flask import request, redirect, render_template
 from app import app, Invitations, Settings, session, Users, htmx, database
 from app.plex import *
+from app.ombi import *
 import secrets
 from app.jellyfin import *
 from app.helpers import *
